backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pydantic==2.5.0
python-multipart==0.0.6
huggingface-hub==0.19.4
pillow==10.1.0
python-dotenv==1.0.0